feat(mobile): jump-to-latest pill in thread view
Scrolling up in a long thread left no quick way back to the newest
reply. Reuse the channel timeline's frosted Latest pill in the thread
detail page: it appears when the newest reply scrolls out of view and
animates back to the tail on tap.

The pill widget is extracted from the channel message list into a
shared public JumpToLatestButton used by both surfaces. _ThreadMessage
and _Avatar move to a part file to keep thread_detail_page.dart under
the 1000-line ceiling.
